Keep tracks from moving in/out of expansion groups
When moving tracks expanded from other tracks in the Oncoprint, only
allow them to move around the other tracks that have the same parent,
effectively making that their track group.

If a track is not the expansion of another group, make the track move
only above its siblings or below their expansion groups and not into
them.