Fix track dragging below bottommost expansion
Due to an oversight in the adjustments to drag tracks over their
siblings' expansions, the visual cues on where the track would end up
would misbehave when dragging a track underneath the expansions of
bottommost track in the group. Integrating the adjustments into the code
used for this scenario fixes the cues.